Emit errors from ansible more cleanly
Sometimes an error comes back from ansible, like on a syntax error or a
missing variable definition or something. These come with failed=True
and a message. Emit those NOT in dict format, but rather with the msg
printed like it's a message.

Also add in a decode step before emitting log lines to catch the places
where we're getting bytes from stdout/stderr from module errors and
other places.

Change-Id: Ic1a589e907327d18aaf3ec1416f76f0bdd261845