Strip zuul work dir from the play banner line
We currently emit the playbook name in the PLAY banner, but it's a full
path because of how we execute ansible-playbook. The zuul work dir is
not interesting information in this context - whereas the canonical path
to the playbook is, such as:

  git.openstack.org/openstack-infra/project-config/playbooks/base/pre

or

  git.openstack.org/openstack-infra/zuul-jobs/playbooks/unittests/pre

Strip the work dir so that the play banner communicates information
interesting to the user. We don't pass it explicitly, so infer it
from the src_root. Additionally, we put playbooks into directories of
the form 'ansible/pre_playbook_0' - so pull that off too.

Change-Id: I2b82b1f6db05a23291260362f0fdb739daf6f743