Strip zuul workdir from the play banner line
We currently emit the playbook name in the PLAY banner, but it's a full
path because of how we execute ansible-playbook. The zuul work_dir is
not interesting information in this context - whereas the canonical path
to the playbook is, such as:

  git.openstack.org/openstack-infra/project-config/playbooks/base/pre

or

  git.openstack.org/openstack-infra/zuul-jobs/playbooks/unittests/pre

Strip the work_dir so that the play banner communicates information
interesting to the user.

Change-Id: I2b82b1f6db05a23291260362f0fdb739daf6f743