Serve html statically
This isn't complete - we need to be able to display the root URL to the
zuul-web process which at the moment neither the executor nor the
scheduler know. So we need to plumb something through, likely in config,
because otherwise folks who are running zuul-web on a sub-url will have
a hard time.

The relative path should be fine though - ../console-stream is
relative to static/stream.html no matter how the two are served.

We'll also want to think about getting tenant names in here, and whether
or not static routes want to be per-tenant or global (could imagine a
per-tenant white-labelling if per-tenant, but also a ton of useless
apache config to deal with one set of static content for a global
install.

Change-Id: I2da7979f448934abe3d41f3a5e50d09004fcccc2