Serve html statically
We need to introduce a few more config variables for this.

The executor needs to know where the zuul-web url is so that it can pass
back status urls with the uuid appended.

Tobias has to run his websocket streamer on a different domain than the
other things, via proxy things - so the url zuul-web is serving for the
static file isn't the same as what it is for the websocket from a
consumer perspective. So introduce a config variable for zuul-web
that allows setting an explicit url for that. If it's not set, use the
relative path from static/stream.html to console-stream as generated by
aiohttp.

There should be no issue with relative paths between static/stream.html
and other paths served by zuul-web, so introduce an info endpoint that
stream.html can hit to grab config options, like websocket_url. Have
that return either the configured or the generated url.

Change-Id: I2da7979f448934abe3d41f3a5e50d09004fcccc2