Update patch set 5
Patch Set 5:

I resolved the earlier issue by causing the implicit role to only be added to jobs and variants *other than* the project-pipeline variants.  I think that's closer to the desired behavior anyway, and causes us to not attempt to change an execution attribute of final jobs.

Patch-set: 5
Label: Verified=0