WIP: Check out implicit branch in timer jobs
So that we may re-use the same jobs for pre and post merge tests,
enqueue an item for every branch of every timer-triggered project
and checkout that branch before running the job.  This means that
rather than having a job for gate plus a job for each stable branch,
we hav just have a single job which runs with different content.

(The old method is still supported using override branches.)

Change-Id: I5c41d2dcbbbd1c17d68074cd7480e6ab83f884ea