Update patch set 2
Patch Set 2:

After learning how badly python degrades with lots (~1000) threads, I'd like to try to keep the number of threads the executor uses to a minimum.  The DiskJobKiller adds another thread for each job we're executing, and it'll mostly be spinning on the condition (I was distraught to learn that's a busy wait in python 2 at least; haven't checked python 3).  But I think we can rework this slightly to avoid the extra thread-per-job.

ExecutorServer.stopJob has all the logic needed to look up a job by its uuid and stop it.  It's written to handle input via gearman, but the guts of the function are simple.  So we could just have the Accountant call a variant of that method directly and cut out the middle man.

The only downside to that is that stopJob calls AnsibleJob.stop() which has a thread.join() in it.  That means that each time we abort a job like this, we will have to wait for the abort to complete.  There are a few ways we can deal with that:

1) Shrug it off.  This probably won't happen that much.  If our response is slow, it probably won't matter too much.  Basically, we'll get good protection against accidental overruns, but poor protection against intentional DOS attacks.

2) Make a variant of AnsibleJob.stop() which doesn't join the thread.  We don't really care in this case -- we can fire and forget.

3) Alter AnsibleJob.stop() to not join the thread.  Even Zuul doesn't care.  It already sent the abort command, the job is dead to it.

4) Create new threads to join the AnsibleJob thread.  This gets us back into lots-o-threads territory, but we'll only create them as needed.  The Accountant would spawn a new thread to handle each kill.  Best case is zero additional threads, worst case we're back to 1 additional thread per job.

After writing all of that out and thinking about it, I like option 3 the best.

Patch-set: 2
Label: Verified=0