Use undefined values instead of 40 zeroes
The git shas 40*'0' are used on the wire protocol to indicate that
the receiving end should create or delete the ref regardless of
its current value.  Both Gerrit and GitHub pass these through
on their event streams unaltered.

To make this more convenient for ansible users, translate these
values into undefined variables.  So that rather than deterimining
that a ref was created by comparing oldrev to
0000000000000000000000000000000000000000 the user can say
"oldrev is not defined".

Change-Id: I1a2a30a0ad53d50325f189a3829f7428db72e1d1