WIP Use yarn and webpack to manage status javascript
This is WIP unti post v3 rollout.

yarn drives package and dependency management. webpack handles
bundling, minification and transpiling down to browser-acceptable
javascript but allows for more modern javascript like import statements.

There are some really neat things in the webpack dev server. CSS
changes, for instance, get applied immediately without a refresh. Other
things, like the jquery plugin do need a refresh, but it's handled just
on a file changing.

If we decide we want to go down this road we should be able to combine
this with the web console streaming and the upcoming console.json
display. We may want to put it into its own repo so that we're not
overlaying javascript/python tools. HOWEVER - there is nothing
inherently blocking keeping it in tree either.

Finally, we can also consider turning the majority of the status page
into a webpack library that other people can depend on as a mechanism
for direct use. Things like that haven't been touched because allowing
folks to poke at the existing known status page without too many changes
using the tools seems like a good way for people to learn/understand the
stack.

Change-Id: I5864e1b2f9952fbc35b08ebdef348c4917564c37