WIP Migrate console streaming to webpack/yarn
This is WIP until post v3 rollout

So that they both work the same, move the streaming web content into the
webpack/yarn structure. Move things so that the built content gets put
into zuul/web/static so that the built-in static serving from zuul-web
will/can serve the files.

Update MANIFEST.in so that if npm run build:dist is run before the
python setup.py sdist, the built html/javascript content will be
included in the source tarball.

Update the zuul job so that the success url contains a source_url
pointing to the live v3 data.

Change-Id: Iab75931ff118ecdf9064ea8671c0fa37d26a1c0a