Update patch set 1
Patch Set 1:

Putting yarn.lock into git was intentional - based on similar thinking to what the rust folks wrote up here: 

http://doc.crates.io/faq.html#why-do-binaries-have-cargolock-in-version-control-but-not-libraries

The thinking in this case is that since this is an app and not a library, storing the exact versions we're working with is advantageous for the same reasons.

That said - this is new territory, and I could certainly see putting yarn.lock into .gitignore instead.

Patch-set: 1
Label: Verified=0