Update patch set 2
Patch Set 2:

I'm working if this is something we could do in a trusted playbook? We should be able to either install ARA or include the callback into a plugins dir? This means it would be an optional dependency for zuul.

One thing to keep an eye is the dependency list ARA will pull in.  

   Successfully installed Flask-0.11.1 Flask-Migrate-2.0.4 Flask-SQLAlchemy-2.2 Flask-Script-2.0.5 Frozen-Flask-0.15 Mako-1.0.7 MarkupSafe-1.0 PrettyTable-0.7.2 PyYAML-3.12 SQLAlchemy-1.1.12 Werkzeug-0.12.2 XStatic-1.0.1 XStatic-Bootstrap-SCSS-3.3.7.1 XStatic-DataTables-1.10.15.1 XStatic-Patternfly-3.21.0.1 XStatic-Patternfly-Bootstrap-Treeview-2.1.3.2 XStatic-jQuery-1.10.2.1 alembic-0.9.3 ansible-2.3.1.0 ara-0.13.3 asn1crypto-0.22.0 bcrypt-3.1.3 cffi-1.10.0 click-6.7 cliff-2.8.0 cmd2-0.7.5 cryptography-2.0.2 debtcollector-1.17.0 decorator-4.1.2 enum34-1.1.6 funcsigs-1.0.2 idna-2.5 ipaddress-1.0.18 itsdangerous-0.24 jinja2-2.9.6 junit-xml-1.7 paramiko-2.2.1 pbr-3.1.1 pyasn1-0.3.1 pycparser-2.18 pycrypto-2.6.1 pyfakefs-3.2 pygments-2.2.0 pynacl-1.1.2 pyparsing-2.2.0 pyperclip-1.5.27 python-dateutil-2.6.1 python-editor-1.0.3 six-1.10.0 stevedore-1.25.0 unicodecsv-0.14.1 wrapt-1.10.10

Patch-set: 2
Reviewer: Gerrit User 4162 <4162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0