Fix github dependent pipeline with merge
When enabling merge in a github gate pipeline merge message formatting
fails with an exception [1]. The reason for this is a missing
initialization of the GithubUser data structure which is using lazy
initialization (presumibly due to api rate limiting). This is fixed by
adding the initialization.

This was uncovered in the tests because the FakeGithubConnection
directly returns a data structure instead of a GithubUser object like
the real GithubConnection. This is fixed by returning a real
GithubUser object linked to a FakeGithub object.

[1] 2017-08-02 07:13:21,323 ERROR zuul.DependentPipelineManager: Exception while reporting:
  Traceback (most recent call last):
    File "/usr/lib/python3.6/site-packages/zuul/manager/__init__.py", line 786, in _reportItem
      ret = self.sendReport(actions, item)
    File "/usr/lib/python3.6/site-packages/zuul/manager/__init__.py", line 168, in sendReport
      ret = reporter.report(item)
    File "/usr/lib/python3.6/site-packages/zuul/driver/github/githubreporter.py", line 73, in report
      self.mergePull(item)
    File "/usr/lib/python3.6/site-packages/zuul/driver/github/githubreporter.py", line 129, in mergePull
      message = self._formatMergeMessage(item.change)
    File "/usr/lib/python3.6/site-packages/zuul/driver/github/githubreporter.py", line 168, in _formatMergeMessage
      if not account:
    File "/usr/lib/python3.6/site-packages/zuul/driver/github/githubconnection.py", line 340, in __len__
      return len(self._data)
  TypeError: object of type 'NoneType' has no len()

Change-Id: Ieb9694ff0c98c8cc4011cd4e36627e8ed647d078