Bindmount /etc/lsb-release into bubblewrap
This like pip use lsb_release, so it is helpful to include this in
bubblewrap.  However, this file doesn't exist on fedora. At the
moment, this means we only now support zuul-bwrap under ubuntu /
debian.

What we should do in a follow up commit, is make this section more
configurable to the end-user. Allowing the user to change these
settings themself, as they see fit.

Change-Id: Ibfed3ace26163da6484966e348e757f7268811f0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>