Update patch set 1
Patch Set 1: Code-Review-1

I think this needs more consideration and discussion. The URI module allows for downloading (to a path) files. That might mean it needs to make use of safe paths. It also can POST data, which could be used to upload (at a provider cost) large amounts of data, creating a DoS of sorts. Granted, this could be done from the node too, but I'd feel better if we discussed it more.

Patch-set: 1
Label: Code-Review=-1