Update patch set 3
Patch Set 3:

Yes, all three options have the same behavior (which is also the current behavior with 'allow-secrets'): the pipeline is assumed to be "pre-review" by default and you have to explicitly indicate otherwise.

The main difference between pre-review and post-review is how that statement looks:

  pre-review: false

or:

  post-review: true

are the same thing, and need to be added to the same pipelines.  I think in my mind, since you have to explicitly declare something to be safe, declaring it to be *true* is more comfortable.

Patch-set: 3