Add some examples to user guide Concepts section
The Zuul Concepts portion of the User Guide introduces a lot of
terminology and concepts that, unless you're already very familiar
with Zuul, remain fairly abstract. This change adds some concrete
examples that help a new user to see the concepts being introduced,
and help tie them all together so that the relationships are, hopefully,
seen between them as well.

Change-Id: I8bdf6e943ac2b79ce213488d070e449d36837d4e