Keep a cache of branches for projects and update on events
github sends events on branch creation. If we get the branch list once
and cache it, then update our local cache on branch creation, we won't
need to get branches every time zuul asks us for them.

We need to look to see about events fired when protection is set/unset,
as that changes how we interact with some branches, and also this needs
to be plumbed into an event so that it can get scheduled since it makes
an outbound call and otherwise could slow-up event ingestion.

Change-Id: I9a16317659921c428c612892c13da5fa533f6872