Add cachecontrol logging to the default config
Most of the logging from cachecontrol is at DEBUG level, so adding it to
the default config should not bother people. Adding it to the default
debug config should help folks debug github webhook issues. Setting it
to INFO will cause the nodaemon logic to set it to debug with the other
things.

Change-Id: Ic76aad08539ea07eae34f4424d94ca24605952ab