Add tests of accessing paths with file is ok
The file module is a normal module so goes through the normal action
plugin. This should validate that the normal action plugin will let a
good path through and block a bad path.

It should also show that the local override check doesn't fail closed,
as it should also allow this to happen.

Change-Id: I5611df4f0448685f2aa509c7d1b5f755f5610c9f