Update patch set 2
Patch Set 2:

Yes, at least that it doesn't break normal operation as in the first patchset I had a bug in there and almost all tests broke.

We might want to add a test case that it also cleans up the cache in the correct way. But I want to make sure that this is still the approach we want to take.

Patch-set: 2