Update patch set 2
Patch Set 2:

So yes, I think a test that ensures the cleanup works is a good idea to include with this change. I've been a bit removed from zuul dev as of late (sadly), but my reading of how tenants and the connections work I would say that this is a good approach.

Patch-set: 2