Update tests to use AF_INET6
zuul_console now properly uses AF_INET6, but some of our tests were
still using AF_INET. There are some sporadic failures with the streaming
test that show some connection errors. It's possible AF_INET6 is a red
herring, but it's also not incorrect to set it, so let's go ahead and
update it.

Change-Id: I278efb939cb6540b5fe2fb1a06495d9be5a7f997