Add test for dependent changes not in a pipeline
When a series of dependent changes get enqueued into a dependent
pipeline that they don't participate in, they are removed from the
pipeline in series.  The first one is removed without incident, but
the second might see the failure of its dependency to merge as a
transitive merge failure.

This used to cause the second change to report, but this commit's
parent fixed that.  Add an explicit test for this situation to verify
and avoid regression.

Change-Id: I85727897c145dab998562dc138affcf08c671593