Be more permissive in configuring nodes
We have a pattern going in many places in the zuul config to allow
a list or a single item, and also to allow a string or a dict. Node
lists in nodeset definitions and in job defitions required that nodes
be specified as a list of dicts. In simple cases, the name of the node
isn't important and people tend to use the label name.

Add the ability for a list of nodes to be given as a list of string or
dict in addition to a list of dicts. If a string is found as one of the
list values it will treated as the value for both label and name of a
Node.

In job definitions, there is already logic that treats a string as a
NodeSet and a list as a list of Nodes. Apply the same new logic to the
list of Nodes case so that strings found in the list continue to be
treated consistently with NodeSet definition.

Change-Id: Ib89be9381a9f0d81ee71db5a0d4675d0cc49e787