Update patch set 1
Patch Set 1: Code-Review-1

I'd like to talk about this a bit more before we decide to do this; this might be the first place where we make a string argument ambiguous (whether it's the nodeset or a simple node def).  I'm not sure how I feel about that.  It means that if anyone mistypes the name of a nodeset, they won't get a syntax error, and it won't be noticed until the job is run.  This may have the unfortunate effect of making the simple case we're trying to solve more difficult for folks, as compared to us just making a set of convenience nodesets in project-config.

If we do decide to proceed with this, this change will need tests and docs.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1