web: add /{tenant}/status route
This change adds a thread to the scheduler to expose the formatStatusJSON
over gearman.

This change also copies the etc/status/public_html files to the zuul-web static.
The zuul.app.js is modified to support the tenant and zuul_web_url parameters.

A follow-up change would remove the duplicated code from etc/ and
webapp module.

TODO:
* add unit-tests

Change-Id: I2238172008a939a56446fbfd6e8621fc16d7e136