Emit job definitions
This does the work to get the needed JJB object associated with the
new Job object, and emits non-useful job stanza. The next patch will
bring actual job content.

If we want to be nice, we should maybe not emit nodes and branches
in the pipeline definitions when we have them in the jobs, but I did
this backwards originally, so that's hard.

Change-Id: I39f7398a9ff965a5ecf68633bcfb210840179d36