Delete IncludeRole object from result object for include_role tasks
The IncludeRole object crashes the callback because ARA tries to
serialize the result._result dict and result._result['include_role']
is not serializable, it's an IncludeRole class instance.

We don't need this key and what is inside so let's just work around
the issue by deleting it for now.

Change-Id: I9216a915e4b6cb7461f30b7b361a75935befc38d
See: https://github.com/ansible/ansible/issues/30385