configloader: don't use path in SourceContext comparaison
The SourceContext is using the in repository yaml path to check if config
statement are part of the same repository. This results in two
zuul.d files context to be differents and some validation fail, for
example when secrets are defined in a dedicated file.

This change removes the path in the SourceContext eq method.

Change-Id: I5500e43faa3cbb7ed470575fe54cb66aed343b9a