Correctly apply irrelevant-files to simple job names
If a job hits the logic as a string, it gets overwritten, but since it's
just a string element of a list it's not a reference member so it does
not update the original list.

Instead, make a new list that we return and then overwrite the old list.
This gets the exclusions added properly.

Also, throw in a set conversion in the middle so that we don't
accumulate multiple copies of the same file exclusions. One copy of each
one is just fine.

Change-Id: Ia8a0d32f3221b252fa093662e6fcd860ba0cabf7