Explicitely print the result status of each playbook run
This will make it easier for people and systems to parse the job
output to find the result of particular playbooks.
For example, in Zuul v2 Elastic-Recheck could rely on the last lines
of the console to reliably tell when the job was completed and what was
it's status.
This new v3 approach allows increased granularity by providing the
ability to filter down to specific playbooks, for example.

Change-Id: I7e34206d7968bf128e140468b9a222ecbce3a8f1