Remove references to pipelines, queues, and layouts on dequeue
Remove these references as soon as items are dequeued so that if
something goes wrong, we don't end up leaking memory.

This shouldn't be an issue in most cases, but was recently when
we had merger processes get stuck running a git command for a merge
job.  The record of the merge job remains until its complete, but
it never will.  The merge job has references to the queue item
and from there to the full layout.

As long as the item continues to be re-enqueued, the outstanding job
is no big deal because pipeline and queue references will be updated,
but if it is dequeued for some reason with an oustanding job, and
that is followed by a reconfiguration, those references will no longer
be updated.  This corrects that by dropping references as soon as
the item is dequeued.

Change-Id: I9dc6e54469fd8dfc8a6d6a70d3c24d0f2261c189