Fix early processing of merge-pending items on reconfig
On a reconfiguration, when we re-enqueue an item, we could end
up creating a job graph for an item before the merge for that
item has returned.  Make sure we use the same methods that the
main processing loop uses so that we maintain the state machine
around waiting for merges and dynamic layouts.

Change-Id: I2b99620cfe1f7666ee3e6297ce041b3f7a02e051