Update patch set 3
Patch Set 3: Workflow-1

This is running into an issue where builds for an item complete, and it is removed from the pipeline by the processing of a previous item, but then we later process the onBuildCompleted event for this item.  Since it no longer has a pipeline associated with it, we can't release the semaphore.

Patch-set: 3
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1