Add a projects_by_name view of zuul.projects
I was in the situation of wanting to know the src_dir of a project for
one of my roles.  It would be much easier to do

 {{ zuul.projects_by_name['project'].src_dir }}

than have to iterate the existing list, try and match the project name
and pull out the src_dir.

This simply proposes adding another view of the projects list, but as
a dict indexed by the project name.

Change-Id: I4476b9d4915d107e29b91229287865bff0ada305