Fix mixed canonical/non-canonical project merge
When defining multiple project stanzas for the same project the
resulting job graphs get merged. This should also work if they are
given as a mixture of canonical and non-canonical project names like
the following stanza.

- project:
    name: review.example.com/org/project1
    check:
      jobs:
        - common-config-job

- project:
    name: org/project1
    check:
      jobs:
        - project1-job

However currently only one of the stanzas are effective depending on
which was processed last.

This can be fixed by canonicalizing the project names when extending
the unparsed config.

Change-Id: Icaf8fca3aa4577b009d691f9a67adcb43ea040f5