Stop storing dependent items on buildsets
Items in a pipeline's queue form a tree (yes, pipeline queues are
actually trees), and when items are deqeued, the links between
them are updated, so that the item now at the root no longer points
to the item that was previously ahead of it.  Normally that would
allow reclamation of all of the memory associated with the item.

However, item buildsets have a list of the dependent items that
were ahead of it at the time the buildest configuration was set.
At this point, these are only used to provide meta-information to
the job about the changes ahead.  Rather than storing references
to the items, only store rendered dictionaries with info about the
changes. That will allow python to reclaim the memory (including
dynamic layouts, and eventually the changes themselves) used by
items as soon as they are dequeued.

Change-Id: I75509a7e9dfeae2337f9f743bec4fcb02ebb903c