Refactor Zuul job to use lightweight Vue components
This patch starts taking parts of the Zuul status UI and converts
them to lightweight Vue components.  At the moment, the UI does
not benefit from improved performance due to the fact that we
still manually render and add things.

However, this will be the start of converting everything into
components in place.  Once the work for this is done, the old
Javascript can be dropped and the components can be used declaratively
directly in the HTML, which will start showing significant speedups.

Change-Id: I5d9007b1f85fb6c54fefc4e86c38d532ddd8d1ba