Add stats for executor and merger count
Report how many executors and mergers are online at any time.  Also
report how many executors are accepting new jobs.  Report the queues
associated with each.

This could be done in the executor or merger clients rather than
centrally in the scheduler, however, determining how many are online
requires a gear admin function, which is relatively expensive.  To
reduce the cost, do it once for the whole system in the scheduler.

The scheduler doesn't directly have a gearman connection (though its
associated rpc listener, merge client, and executor clients do).
It didn't seem necessary to add another client for this, and the
rpc listener seemed the most appropriate connection to borrow (it's
purpose is to expose scheduler functions over gearman).  The method
to count functions is added to it, and it in turn is now started
directly from the scheduler.

Change-Id: I09659c29431ebac7ecd1869cc4c1356026c03d26