Make multiple parent jobs returning same var undefined behavior
The test case test_data_return currently has a race. It has a child
job with two parents returning the same variable back. Currently it
expects that there is a defined ordering in which the data is
forwarded to the child job which is currently not guaranteed.

One option to fix this is to drop this guarantee and fix the test to
not leverage it.

Change-Id: If8075ade11defbf8358c06f38dd3b14a62a01aff