Use end time based sorting for applying parent data
The test case test_data_return currently has a race. It has a child
job with two parents returning the same variable back. Currently it
expects that there is a defined ordering in which the data is
forwarded to the child job which is currently not guaranteed.

This can be fixed by changing the variable apply order to be time
based by build end time. This might be the closest to the expected
user behavior and makes it possible to easily overwrite vars downwards
the dependency tree. However this introduces a risk of race conditions
as noted in the docs when used without caution.

Change-Id: Ib811915cf62822e3ad0efa1acd9f999bffd85136