Update patch set 2
Patch Set 2:

I also thought about just using name but I want to make regex matching a clear decision as this is much more expensive than the default name matching.

If we want that we might want to mark this somehow that it's a regex like strings starting with 'r/' or something like that or maybe adding another bool field for activating regex matching.

Patch-set: 2
Label: Verified=0