Update patch set 1
Patch Set 1:

The real fix for this is to allow nulls on those columns, and set it to null in these cases.

Do we want to jump straight to that, or merge this as a stop-gap (but with a FIXME)?

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0