Update patch set 1
Patch Set 1:

I'd rather do a more correct fix here, but am not entirely clear on how to tell sqlalchemy to allow nulls in integer columns. Is that being enforced by some prophylactic validation in the ORM or does it require changes to the underlying table schema in MySQL (and will that in turn need a migration)?

Patch-set: 1