Update patch set 1
Patch Set 1: Code-Review+2

This screams out for a test, though it would require some monkeypatching to really make deterministic. Also, I wonder if we should consider passing "ignore" to our decode() calls. If something is able to slip binary data through to the log, I'd rather we keep running and drop it or replace it with a weird thing than explode.

Patch-set: 1
Label: Code-Review=+2