Do not add invalid projets via the /keys API
If someone requests a project that does not exist in the layout, the
old version of _handle_keys would add that project to the layout, and
then fail with a 500 because the public_key attribute doesn't exist.

Instead we should use the readonly version of the method so that a
malicious attacker cannot cause us to exhaust all our RAM by adding tons
of erroneous project names.

Change-Id: I3ace2d792542dabd9a240c38274c307a11afe2b3