Fix gerrit branch creation detection
We detect when a new branch is created so that we know to perform
a reconfiguration which will include the new branch.  However,
the detection for this case in the Gerrit driver still had the
pre-2.13 form of branch representation.  Update it to support both
forms, and update the tests to use the new form.

Change-Id: I0480482a59a9f289ca4e449ce70fd1c0872e7999