Make enqueue-ref <new|old>rev optional
As described in the documentation in
Ibd7ee306bc461d1597c9c8febcaad89a48b9ff96 the new/old rev arguments
aren't required in all situations.  For example when triggering a
periodic job and pretending to be the timer event, you just need the
refs/head/branch.  For a release/tag replay, I believe you only need
the newref, so setting them both to zero by default triggers the check
incorrectly.

Thus we set the arguments to None intially and only if they are
explicitly set do we check they're not the same.  If they're unset by
the user, we just set them to zero for the rpc calls.

Change-Id: I8125010efcf8d26dc3a99c87fe2c945314072b72