Update patch set 20
Patch Set 20: Code-Review-1

(1 comment)

I'm not sure about the approach of querying very build from every database connection. Unless the connections are set up very carefully this could result in a lot of duplicates.

My opinion would be to have a config variable for [web] which selects which sql connection to use for rendering results. It's then expected that that connection is used as a reporter for storing all of the results you wish to appear in the dashboard.

Happy to be convinced otherwise.

Patch-set: 20
Label: Code-Review=-1