Allow run to be list of playbooks
Like pre-run and post-run, allow a user to run a list of playbooks for
a job. One example would be your job workflow would be to run multiple
playbooks on the same host but use different groups in an inventory
file. This would allow for a single node from nodepool, but allow for
multiple SSH attempts (helping with ansible variable namespacing).

Change-Id: I56b9c01a2d8bc7ce433439dfa6b0e73657d5ecc8
Signed-off-by: Paul Belanger <pabelanger@redhat.com>